Avoid int type warning

+++ onyx/onyx/draw/model.cpp
@@ -75,7 +75,8 @@ void Model<D>::Draw(const VkCommandBuffer p_CommandBuffer, const u32 p_InstanceC
                     const u32 p_FirstVertex) const noexcept
 {
     TKIT_ASSERT(!m_IndexBuffer, "[ONYX] Model does not have indices, use Draw instead");
-    vkCmdDraw(p_CommandBuffer, m_VertexBuffer.GetInfo().InstanceCount, p_InstanceCount, p_FirstVertex, p_FirstInstance);
+    vkCmdDraw(p_CommandBuffer, static_cast<u32>(m_VertexBuffer.GetInfo().InstanceCount), p_InstanceCount, p_FirstVertex,
+              p_FirstInstance);
 }
 
 template <Dimension D>
@@ -83,8 +84,8 @@ void Model<D>::DrawIndexed(const VkCommandBuffer p_CommandBuffer, const u32 p_In
                            const u32 p_FirstIndex, const u32 p_VertexOffset) const noexcept
 {
     TKIT_ASSERT(m_IndexBuffer, "[ONYX] Model has indices, use DrawIndexed instead");
-    vkCmdDrawIndexed(p_CommandBuffer, m_IndexBuffer.GetInfo().InstanceCount, p_InstanceCount, p_FirstIndex,
-                     p_VertexOffset, p_FirstInstance);
+    vkCmdDrawIndexed(p_CommandBuffer, static_cast<u32>(m_IndexBuffer.GetInfo().InstanceCount), p_InstanceCount,
+                     p_FirstIndex, p_VertexOffset, p_FirstInstance);
 }
 
 template <Dimension D> const VertexBuffer<D> &Model<D>::GetVertexBuffer() const noexcept
